Refactor navigation and header search components
